Payment criada

+++ Projetos/spring_mte/src/main/java/com/mte2023/spring_mte/resources/UserResource.java
@@ -3,9 +3,6 @@ package com.mte2023.spring_mte.resources;
 import java.util.List;
 import java.util.Optional;
 
-import com.mte2023.spring_mte.exceptions_generic.BadRequestException;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,11 +10,17 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
 import com.mte2023.spring_mte.entities.User;
+import com.mte2023.spring_mte.exceptions_generic.BadRequestException;
 import com.mte2023.spring_mte.services.UserService;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 
 @RestController
 @RequestMapping(path = "/users", produces = "application/json;charset=UTF-8")
@@ -64,12 +67,12 @@ public class UserResource {
         }
         throw new BadRequestException("Erro ao deletar o usuário.", 400);
     }
-/*
+
     @PutMapping(path = "/{id}")
-    public ResponseEntity<User> update(@PathVariable Long id, @RequestBody User obj){
+    public ResponseEntity<User> update(@PathVariable Integer id, @RequestBody User obj){
         obj = userService.update(id, obj);
         return ResponseEntity.ok().body(obj);
     }
 
-     */
+     
 }
